12-web/pyproject.toml
[project]
name = "12-web"
version = "0.1.0"
description = "web"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-index-embeddings-ollama>=0.6.0",
    "llama-index-llms-ollama>=0.6.2",
    "llama-index-node-parser-topic>=0.2.0",
    "llama-index-readers-web>=0.4.4",
    "llama-index-vector-stores-qdrant>=0.6.1",
]
